+++ assets/javascripts/kitten/components/links/link-list.js
@@ -14,27 +14,20 @@ export class LinkList extends Component {
   }
 
   renderItem(element) {
-    const {
-      key,
-      item,
-      href,
-      active,
-      weightItem,
-      className,
-      ...others
-    } = element
+    const { key, item, href, active, weight, className, ...others } = element
 
-    const { color, lineHeight, itemMargin, weight } = this.props
+    const { color, lineHeight, itemMargin } = this.props
 
     const linkListClassName = classNames('k-LinkList__link', className, {
       'is-active': active,
       'k-LinkList__link--light': color == 'light',
       'k-LinkList__link--dark': color == 'dark',
       'k-LinkList__link--normalLineHeight': lineHeight == 'normal',
-      'k-LinkList__link--regularWeight': weight == 'regular',
-      'k-LinkList__link--lightWeight': weight == 'light',
-      'k-LinkList__item--regularWeight': weightItem == 'regular',
-      'k-LinkList__item--lightWeight': weightItem == 'light',
+      'k-LinkList__link--regularWeight':
+        this.props.weight == 'regular' && !weight,
+      'k-LinkList__link--lightWeight': this.props.weight == 'light' && !weight,
+      'k-LinkList__item--regularWeight': weight == 'regular',
+      'k-LinkList__item--lightWeight': weight == 'light',
     })
 
     const linkListItemClassName = classNames('k-LinkList__item', {
@@ -82,7 +75,7 @@ LinkList.defaultProps = {
 LinkList.defaultProps = {
   className: null,
   margin: true,
-  items: [], // Eg: [{ key: …, item: …, href: … }]
+  items: [], // Eg: [{ key: …, item: …, href: …, weight: …, }]
   color: 'dark',
   lineHeight: null,
   itemMargin: null,